Debug and test enhancements
Enhance dbg-vec function so that it supports debug printing of
internal structure of transient versions of PersistentVector and RRB
vector, and even the Clojure APersistentVector$SubVector type.

Add utility function all-vector-tree-nodes and a few other
property-checking functions that verify some conditions that should be
true for the internal structure of all vector collections.

basic-node-errors checks conditions that should be true for all
Clojure vector types covered here, including Clojure peristent
vectors, RRB vectors, and their transient versions.

ranges-not-int-array is specific to RRB vectors as they are
implemented by core.rrb-vector.  It can detect ranges in slot 32 of a
node array that are not a Java array of int's.  There is an open bug
that leaves an RRB vector such that it violates this condition, and
this function can detect such bad RRB vectors.

Add a new utility protocol PTransientDebugAccess that is only for
debug purposes, reading private fields of the Transient deftype for
printing via dbg-vec.

Add a trivial function dbgln that is simply println, but its
definition can easily be changed in one place to make it a no-op
instead of printing.  Ideally it should be a macro, to even avoid the
function call, and the caller calculating its arguments.

Add some code to test namespace clojure.core.rrb-failing-tests that
wraps basic vector functions, so that they check their return values
for a few kinds of errors before returning, for faster catching of
problems.

TBD: It would be good to write some small utility code similar to
spec's instrument, or the memoize function, that wraps more such
functions in a way that requires less code, and hopefully works for
functions that return vectors in clojure.core, too.