Add option of aborting more quickly to rrb-failing-tests
It requires a one line manual edit, because it is only intended right
now for interactive testing.

Slightly simplify the puzzle-b test function so that the vector
elements are integers, rather than 2-element vectors.  The problem
still occurs, because it is due to the structure of the data structure
representing the top level vector, not what its elements are (as are
nearly all problems with the vector implementation).