Several enhancements to Clojure debug namespace
Make one of the invariant checker error messages show the full
sequence, rather than a less informative LazySeq object.

Add abbrev-type-name utility function for debug purposes.

Enhance dbg-vec to show the tidx value for transient vectors, the
number of valid array elements in the tail, because the array is
always 32 elements long for transients, so it is not always obvious
how many of them are valid.

Minor fix to a debug function in case the shift value is not a
multiple of 5, but that would be an error if it did, anyway, one that
is already checked for in basic-node-errors fn.

Add fraction-full debug function to estimate the fraction of the full
tree capacity of elements is used by vector elements.  Useful for
knowing whether the tree is getting 'tall and skinny'.

Rename a few debug functions to have slightly better names.