Add new dbg-* versions of catvec, subvec, slicev, splicev, splice-rrbts
They do the extra sanity checks that were implemented and used
earlier, plus the checks that verify that the return values are the
correct functions of their inputs.

Changed all tests in rrb-vector-common-test to use these new
functions.

There is still a fair amount of experimental code that I should
probably remove from the debug namespace, but I want to check in
something that is working right now, and clean up in a later commit.