CRRBV-14 workaround fix - create perfectly balanced tree if needed
The ideal fix for this issue would be to change splice-rrbts so that it
never returns a tall skinny tree, as it currently does in some cases.

This commit represents a workaround, where instead of the ideal fix we
quickly detect if the normal splice-rrbts return value would be too
tall and skinny, and if so, create a new perfectly balanced tree in
O(N) time, where N is the number of elements.

In general, this will not be good for preserving the desired run time
of the catvec operations.  However, at least for the test case we have
that leads to the issue of CRRBV-14, out of over a thousand catvec
operations, only one of them triggers this slower fallback action.  It
remains to be seen whether there are cases that this can happen more
often.

In the mean time, though, this does prevent the crash seen from the
tree getting so deep that shift goes over 64.