CRRBV-22 fix for ClojureScript for off by one bug near 32^2 elements
The change here is motivated by the corresponding code in Clojure's
PersistentVector class, both for the persistent and transient
versions, in its popTail methods.  I believe one way to think about
why it is (cnt - 2) instead of (cnt - 1) here is that cnt is the
number of elements in the factor before it is pop'd, so (cnt - 1) is
the new number of elements after popping.  (cnt - 2) is the index of
the new last element after popping.