Change error about tree deeper than needed to a warning
and also disable it when it occurs as the result of a slicev call.  I
have found it to be fairly common for slicev's return value to trigger
this condition, and as far as I can tell it is not an error, simply a
design choice in how slicev is implemented, with the assumption that
if that value is ever spliced with another vector, all invariants
should be restored in the result of the splicev operation.