Add deps.edn aliases and jdo/sdo shortcuts for collection-check tests
Change suffix of source file using collection-check to .cljc, so it
can be used from both Clojure and ClojureScript.

Add some aliases to deps.edn for running collection-check tests, and
task names to the jdo and sdo shell scripts for running them.

Show the Java version (if running Clojure) and Clojure/ClojureScript
version when running the first deftest of a collection of tests, and
show the elapsed time when each one completes.