Perhaps an actual good fix for CRRBV-17
I used the dbg-vec function to look at many trees produced as
intermediate results in the test-splicing deftest, and noticed that
several of them had an inefficient structure where the top two levels
of the tree had so few grandchildren that all of those grandchildren
could be made direct children of the root, and the rest of the tree
structure would be fine, and the tree would be one level shallower as
a result.

It might be that there is a simple fix to the existing splice-rrbts
function that would avoid this kind of tree being created in the first
place, but this commit represents an experiment where I created a
"peephole optimizer" that at the end of running the normal
splice-rrbts function, checks whether the resulting tree could be
reduced one level in depth by this technique.  If so, it does it,
otherwise it returns the tree it would have without these changes.

The previous O(N) "fallback" code is still there, to avoid crashes if
the peephole optimizer does not produce good enough results.

I was pleasantly surprised that for all of the existing tests,
including the one for CRRBV-17, the fallback code never runs, and the
tree never gets too deep.  For the CRRBV-17 test case, the peephole
optimizer actually makes improvements for about half of the 100,000
catvec calls, but that should not harm the O(log N) run time, because
the peephole optimizer runs in O(log N) time worst case -- often it is
O(1) time, but in some cases with regular subtrees it might need to
traverse the full tree depth in order to count the number of vector
elements in each of those subtrees.