Changed most dbg-* functions to checking-* in core.rrb-vector.debug
Before I made any of these additions, there was already a function
dbg-vec whose only purpose was to print out the internals of a vector
tree data structure.  Having that function name and also dbg-catvec,
where dbg-catvec is just like catvec except it performs extra checks
on its return value, is simply a poorly chosen name.  checking-catvec
is not much longer, and the name is enough different that it is not
easily confused with the difference between vec and dbg-vec.